+++ src/engine/bench.rs
@@ -5,9 +5,9 @@ use std::time::Instant;
 use anyhow::Result;
 
 use crate::chess::position::fen::Fen;
-use crate::limits::Limits;
-use crate::search::SearchManager;
-use crate::tt::Table;
+use crate::engine::limits::Limits;
+use crate::engine::search::SearchManager;
+use crate::engine::tt::Table;
 
 const BENCHMARK_FENS: [&str; 50] = [
     "r3k2r/2pb1ppp/2pp1q2/p7/1nP1B3/1P2P3/P2N1PPP/R2QK2R w KQkq a6 0 14",
