+++ src/engine/eval.rs
@@ -1,10 +1,10 @@
 use crate::chess::position::Position;
 use crate::chess::{Color, Role, Square};
-use crate::search;
+use crate::engine::search::MAX_PLY;
 
 pub const INFINITY: i16 = 32_001;
 pub const MATE: i16 = 32_000;
-pub const MATE_IN_PLY: i16 = MATE - search::MAX_PLY as i16;
+pub const MATE_IN_PLY: i16 = MATE - MAX_PLY as i16;
 pub const DRAW: i16 = 0;
 
 impl Position {
