+++ src/lib.rs
@@ -1,19 +1,12 @@
-pub mod bench;
 pub mod chess;
-pub mod limits;
-pub mod search;
-pub mod tt;
-pub mod uci;
+pub mod engine;
 
 #[cfg(feature = "datagen")]
 pub mod datagen;
 
-mod eval;
-mod movepicker;
-mod utils;
-
 use crate::chess::movegen;
 use crate::chess::position::zobrist;
+use crate::engine::search;
 
 pub fn init() {
     movegen::init_tables();
